config.php clean up
