+++ lib/extras.php
@@ -2,6 +2,28 @@
 
 namespace Roots\Sage\Extras;
 
+use Roots\Sage\Config;
+
+/**
+ * Add <body> classes
+ */
+function body_class($classes) {
+  // Add page slug if it doesn't exist
+  if (is_single() || is_page() && !is_front_page()) {
+    if (!in_array(basename(get_permalink()), $classes)) {
+      $classes[] = basename(get_permalink());
+    }
+  }
+
+  // Add class if sidebar is active
+  if (Config\display_sidebar()) {
+    $classes[] = 'sidebar-primary';
+  }
+
+  return $classes;
+}
+add_filter('body_class', __NAMESPACE__ . '\\body_class');
+
 /**
  * Clean up the_excerpt()
  */
